backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Job Tracker application"
authors = [{ name = "Rohit Sharma", email = "dev.apexinsideapex@proton.me" }]
dependencies = [
  "fastapi>=0.104.1",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.5.0",
  "python-multipart>=0.0.6",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-decouple>=3.8",
  "psycopg2-binary>=2.9.9",
  "sqlalchemy>=2.0.23",
  "alembic>=1.13.1",
  "pyresparser>=1.0.6",
  "nltk>=3.8",
  "pdfplumber>=0.9.0",
  "python-docx>=0.8.11",
  "pip>=25.1.1",
  "spacy>=3.8.7",
]

requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "pytest>=7.0.0",
  "httpx>=0.24.0",
  "black>=23.0.0",
  "flake8>=6.0.0",
]
